requirements.txt
flask>=3.0
python-dotenv>=1.0
markdown>=3.5
pygments>=2.17
